feat: moved useElements hook to own file
