+++ app/hooks.ts
@@ -14,26 +14,9 @@ import {
   RIGHT_TEXT
 } from "./constants"
 import useDebounce from "./hooks/useDebounce"
+import useElements from "./hooks/useElements"
 import useFonts from "./hooks/useFonts"
 import type { SelectorProps, StylesData } from "./types"
-
-export const useElements = () => {
-  const widgetBody = useRef(document.querySelector<HTMLDivElement>(GOAL_WIDGET))
-  const image = useRef(document.querySelector<HTMLImageElement>(GW_IMAGE))
-  const leftText = useRef(document.querySelector<HTMLDivElement>(LEFT_TEXT))
-  const rightText = useRef(document.querySelector<HTMLDivElement>(RIGHT_TEXT))
-  const progressBar = useRef(
-    document.querySelector<HTMLDivElement>(GW_PROGRESS_BAR)
-  )
-
-  return {
-    [GOAL_WIDGET]: widgetBody.current,
-    [GW_IMAGE]: image.current,
-    [LEFT_TEXT]: leftText.current,
-    [RIGHT_TEXT]: rightText.current,
-    [GW_PROGRESS_BAR]: progressBar.current
-  }
-}
 
 export const useDefaultStyles = () => {
   const elements = useElements()
