+++ app/hooks/useElements.ts
@@ -1,0 +1,29 @@
+import { useRef } from "react"
+
+import {
+  GOAL_WIDGET,
+  GW_IMAGE,
+  GW_PROGRESS_BAR,
+  LEFT_TEXT,
+  RIGHT_TEXT
+} from "~app/constants"
+
+export const useElements = () => {
+  const widgetBody = useRef(document.querySelector<HTMLDivElement>(GOAL_WIDGET))
+  const image = useRef(document.querySelector<HTMLImageElement>(GW_IMAGE))
+  const leftText = useRef(document.querySelector<HTMLDivElement>(LEFT_TEXT))
+  const rightText = useRef(document.querySelector<HTMLDivElement>(RIGHT_TEXT))
+  const progressBar = useRef(
+    document.querySelector<HTMLDivElement>(GW_PROGRESS_BAR)
+  )
+
+  return {
+    [GOAL_WIDGET]: widgetBody.current,
+    [GW_IMAGE]: image.current,
+    [LEFT_TEXT]: leftText.current,
+    [RIGHT_TEXT]: rightText.current,
+    [GW_PROGRESS_BAR]: progressBar.current
+  }
+}
+
+export default useElements
